Preserve participant identity when updating nested participants
Once a participant is created alongside an appearance, its "identity" fields (given name, family name and birthdate) cannot be changed by editing a registration. Instead, the appearance needs to be deleted and created with a new participant. This is important so we can safely assign existing participants to new registrations, without running the risk of someone accidentally switching their identity _everywhere_ by editing a registration. Example: Someone wants to swap out a participant for another, but simply changes the former participant's name and birthdate to the next's in the form. This would break because the form field still holds the former participant's id, essentially swapping that participant's identity in all their performances, which is not the desired result.)