Fix changeset flattening logic by using defaultsDeep
Without this change, values below the top hierarchy level that were not contained in params would be omitted from the result. This becomes apparent e.g. when disabling certain form fields, causing the param to be omitted altogether. Before this change, this would cause the form field to be empty despite there being an existing value in the changeset data.