"Reuse" existing participants in new appearances if they match
This is done by looking at the to-be-inserted participants' identity fields (given_name, family_name and birthdate) and looking for matches in the database. If a match is found, that participant's changeset is converted to an :update changeset, so that non-identity fields (e.g. phone number) are updated to the latest information across the DB.