Prepopulate Kimu form in controller if the contest has only one category
Most logic for populating the registration form lives in JS code, but if a Kimu contest has only one category (as typically is the case), we can already prepopulate the performance struct in the controller itself. This means that the category will already be selected when the user opens the form.