Change type of Performance stage time from DateTime to NaiveDateTime
This is semantically more accurate, since performances are thought to happen at a certain wall time, regardless of time zone changes. With this change, all stage times are transmitted and exposed in the app without a timezone (e.g. "2019-01-01T09:00:00"). In case we need to relate them to absolute moments in time later, we can use the contest's time zone to calculate the UTC timestamp. The ambiguities caused by DST should not matter here, because such shifts take place at night and performances don't.