Exclude unused stages when loading public contests
This is achieved by joining the contest's stages (via host) and performances (via contest categories), and requiring performances to be staged on one of the joined stages. Next, the joined stages are assigned manually during the preload (`stages: s`). This way, stages without performances in that contest won't be preloaded.