Remove <wohnbundesland> field from exported XML
Based on JMDaten documentation, the possible field values are DSN, DSWM and DSOM – names of regions which strictly speaking no longer exist since the decision to migrate to more flexible "groupings" (Verbund) that are no longer based on geography and whose host memberships may change over the years. We therefore opt to omit the field in the export and leave the decision about what value to select in JMDaten to the person who imports the file.