Track and use predecessor host instead of contest
Since there can be performances whose predecessor contest is not in the database (but who still need some kind of predecessor info), we record the predecessor host alongside the contest. This also optimizes the code a bit, as we previously accessed predecessor contests only to retrieve their associated host, which can now be reached with a single join instead of two. However, if available we want to keep the predecessor contest linked as well, to prepare for future scenarios such as nested navigation to a predecessor performance (which requires the contest ID).