Validate predecessor fields on Performance for round 2
If a 2nd-round performance already has associated predecessor data (`predecessor` or `predecessor_contest`), we don't allow the predecessor host to be user-edited, because this would probably render it inconsistent with the above fields which cannot be changed by the user. If no predecessor data exists, we require the `predecessor_host_id` to be present.

For performances in round 0 and 1 the `predecessor_host_id` attribute is simply ignored during casting.