Fix age groups after merging or editing participant
So far we only recalculated the age groups when the entire performance was edited (meaning we were operating on a Changeset). This led to stale age groups when a participant’s birthdate was edited in isolation, or two participants with different birthdates were merged. To fix this, we feed the entire Performance to AgeGroupCalculator and receive a changeset that fixes the age groups.