Add WESPE nomination support to certificates
In some categories (e.g. “Jumu open”), receiving a 1st prize may lead to a WESPE nomination which should be printed on the certificate. However, since this only applies to certain contests (currently only LW level), we need to add a flag to ContestCategory that can be activated where needed.