Replace Phauxth by phx.gen.auth code
This gives us full control over auth code in the future and removes a big external dependency, but also means we need to keep it up-to-date as per Phoenix upgrade notes.

Some features of phx.gen.auth, such as registration and email confirmation, were skipped at this time as we don't need them (yet).