Consider WESPE nominations “advancing”
This currently applies to performances in the “Jumu open” category, which so far weren’t treated as advancements by the software but instead got separate logic to determine whether to show the nomitation text on the certificate. However, the BW team wants them to be part of the advancement export file, so we tweak the logic slightly.